Home

PressureFieldContact

This module implements the elastic foundation-themed contact model for rigid body dynamics described in this short video. This paper describes the method in greater detail.

Design

Words here

bullet point?

another?

Installation

To install you

PressureFieldContact.jl Documentation

Polygon clipping

Geometry

Basic Shapes

Swept Meshes

Rotationally Symmetric Meshes

Quick-Start

Friction

Add friction

PressureFieldContact

Design

Installation

PressureFieldContact.jl Documentation

Functions

Index

Algorithms

Functions

PressureFieldContact.calcXd! — Function.

calcXd!(xx, x, m, tm, t)

Conventions: n̂ refers to the contact surface normal that points into body B v_cart refers to + velocity of B - the velocity of A the wrench is the wrench applied TO body A

source

Index

PressureFieldContact.Geometry.eMesh

PressureFieldContact.Geometry.output_eMesh_box

PressureFieldContact.Geometry.output_eMesh_half_plane

PressureFieldContact.Geometry.output_eMesh_sphere

PressureFieldContact.NumericalTricks.centroid

PressureFieldContact.NumericalTricks.mul_then_un_pad

PressureFieldContact.NumericalTricks.one_pad_then_mul

PressureFieldContact.Tri_Tet_Intersections.clip

PressureFieldContact.Tri_Tet_Intersections.clip_in_tet_coordinates

PressureFieldContact.Tri_Tet_Intersections.clip_plane_tet

PressureFieldContact.Tri_Tet_Intersections.cut_clip

PressureFieldContact.Tri_Tet_Intersections.getTetQuadRule

PressureFieldContact.Tri_Tet_Intersections.getTriQuadRule

PressureFieldContact.Tri_Tet_Intersections.zero_small_coordinates

PressureFieldContact.add_friction_bristle!

PressureFieldContact.add_friction_regularize!

PressureFieldContact.calcXd!