Geometry

Geometry

This contact implementation represents rigid objects with triangular meshes and compliant objects with tetrahedral meshes. This package uses the custom geometry data type eMesh. Convenience functions to create meshes for common shapes are documented below.

PressureFieldContact.Geometry.eMesh — Type.

struct eMesh{T1<:Union{Nothing, Tri}, T2<:Union{Nothing, Tet}}

Data structure for holding geometry data.

source

Basic Shapes

PressureFieldContact.Geometry.output_eMesh_box — Function.

output_eMesh_box()
output_eMesh_box(r)
output_eMesh_box(r, c)

Outputs an eMesh for a box.

source

PressureFieldContact.Geometry.output_eMesh_sphere — Function.

output_eMesh_sphere()
output_eMesh_sphere(rad)
output_eMesh_sphere(rad, n_div)

Outputs an eMesh for a sphere. Larger values of n_div create finer discretizations.

source

PressureFieldContact.Geometry.output_eMesh_half_plane — Function.

output_eMesh_half_plane()
output_eMesh_half_plane(plane_w)
output_eMesh_half_plane(plane_w, is_include_vis_sides)

Outputs an eMesh for a half-plane.

source

Swept Meshes

Rotationally Symmetric Meshes

Mesh from an STL file

For an example of this this example: