Home

PressureFieldContact

This module implements the elastic foundation-themed contact model for rigid body dynamics described in this short video. This paper describes the method in greater detail.

Design

Words here

bullet point?

another?

Installation

To install you

PressureFieldContact.jl Documentation

Geometry

Functions

Quick-Start

Friction

PressureFieldContact

Design

Installation

PressureFieldContact.jl Documentation

Functions

Index

Algorithms

Functions

PressureFieldContact.calcXd! — Function.

calcXd!(xx, x, m, tm, t)

Conventions: n̂ refers to the contact surface normal that points into body B v_cart refers to + velocity of B - the velocity of A the wrench is the wrench applied TO body A

source

Index

PressureFieldContact.Binary_BB_Trees.output_eMesh_box

PressureFieldContact.Binary_BB_Trees.output_eMesh_sphere

PressureFieldContact.calcXd!