Home

PressureFieldContact

PressureFieldContact implements...

Design

Words here

bullet point?

another?

Installation

To install you

PressureFieldContact.jl Documentation

PressureFieldContact

Design

Installation

PressureFieldContact.jl Documentation

Functions

Index

Algorithms

Functions

PressureFieldContact.calcXd! — Function.

calcXd!(xx, x, m, tm, t)

Conventions: n̂ refers to the contact surface normal that points into body B v_cart refers to + velocity of B - the velocity of A the wrench is the wrench applied TO body A

source

Index

PressureFieldContact.calcXd!