Geometry

Geometry

This contact implementation represents rigid objects with triangular meshes and compliant objects with tetrahedral meshes.

This package uses the custom geometry data type eMesh.

Convince functions to create meshes for common shapes are documented below.

Functions

PressureFieldContact.Geometry.output_eMesh_box — Function.

output_eMesh_box(radius, center)

Outputs an eMesh for a box. Example: meshbox = outputeMesh_box(0.05)

source

PressureFieldContact.Geometry.output_eMesh_sphere — Function.

output_eMesh_sphere(radius, n_div::Int64)

Outputs an eMesh for a sphere. Larger values of ndiv create finer discretizations. Example: meshsphere = outputeMeshsphere(0.05)

source

Missing docstring.

Missing docstring for output_eMesh_half_plane. Check Documenter's build log for details.